Allow charset to have an extra ; at the end
Some websites have an extra semicolon at the end of their Content-Type
header, after the charset parameter, which causes decoding content to
fail. A brief reading of the HTTP 1.1 spec implies that this is not
explicitly permitted, but the web is not always up to spec. P6's
LWP::Simple also allows this.