::Message to decode for .get; add decode fallback [fixes #35]
HTTP::Message and HTTP::UserAgent had the same decode logic, so
I removed it from UserAgent and used the $request object's message
object's decode-content method. I *think* this was an artifact from
when the content was stored in a my $content instead of the
$response.content attribute.

The decode-content logic was changed to use a saner default of utf-8
unless content-type contains 'text/' (which uses ascii by default
per rfc). However, an additional step should eventually be added
for when the charset is missing from the header; it should attempt
to extract the charset from the $response body meta tag attributes
(chicken and egg problem?). Issue #35 covers this: baidu.com does
not send a charset in the header, but it *is* in the html's meta tag.