Connection default to explicit close
"In HTTP 1.1, all connections are considered persistent unless declared otherwise."

`HTTP::UserAgent` relies on the connection being closed to determine when its done receiving when no `Content-Length` or `Transfer-Encoding` headers are sent. But an HTTP/1.1 server may default to keeping the connection open after its sent its payload if you do not tell it otherwise. So default to `Connection: close` if not manually set in the request itself.

Resolves #105