Do not use HTTP_PROXY if this in the CGI
This is a pre-emptive fix for a potential vulnerability to
https://www.kb.cert.org/vuls/id/797896, though right now there
is no actual CGI implementation for Perl 6 that I know of.

If the REQUEST_METHOD environment variable is set, then ignore
HTTP_PROXY environment variable, but use CGI_HTTP_PROXY instead.

This fixes #148 though a more refined handling of the proxy thing
is probably required.