Save memory by not storing full request history
See #169.
We used to store the full history of the response, just to count redirects.
This used up lots of RAM for very little gain.

This commit replaces it by a simple integer that stores the count of
redirects in a row.

Since @.history was a public attribute, this changes the public API,
and as such is subject to discussion.