Connection error 404 Bad Request via proxy
Everytime I tried to conect via a proxy, I was getting a `400 Bad
Request` error. Debugging the problem, I found out that the `Str` method
of `HTTP::Request` was ignoring the case where the `$.file` field was a
copy of the `$.url` field and, because of it, started with `"http"`. The
connection was asking my proxy for a `GET /http://etcetc`