Reduce scope of SSL lock
A lock was added in #193 (45fdcf359d) to fix an issue resulting in
a segmentation fault being triggered when creating IO::Socket::SSL
objects in an async process. At the time, the reason for the error
was assumed to be in IO::Socket::SSL or in one of its underlying
dependencies.

The lock added protected the entire block where IO::Socket::SSL was
conditionally loaded and a new socket was constructed. While this
solved the problem, it meant that while using HTTPS, HTTP::UserAgent
was fundamentally incompatible with async environments.

However, the issue does not seem to be related to IO::Socket::SSL
but to the conditional loading logic, and reducing the scope of the
lock not only continues to protect against the error, but results in
much better support for async requests.