Decode header as ISO-8859-1
Before this change, requests to servers that send UTF-8 data (or any
non-ASCII data) as a header value (such as the `x-duck` that
http://icanhazip.com sets to 🦆) would be unparseable, and die with

> Will not decode invalid ASCII (code point (-16) < 0 found)

error.

Although servers _shouldn't_ be sending ducks in HTTP/1.1 headers,
clients also _shouldn't_ die when trying to parse those responses.

On this, [RFC 7230 § 3.2.4] states:

> Historically, HTTP has allowed field content with text in the
> ISO-8859-1 charset [ISO-8859-1], supporting other charsets only
> through use of [RFC2047] encoding.  In practice, most HTTP header
> field values use only a subset of the US-ASCII charset [USASCII].
> Newly defined header fields SHOULD limit their field values to
> US-ASCII octets.  A recipient SHOULD treat other octets in field
> content (obs-text) as opaque data.

It is not clear what the spec suggests clients should do regarding
this 'opaque data', but the solution in this patch means that
responses can be parsed, and users that _do_ want to recover whatever
opaque data was sent in a header can do so by re-encoding the header
value and decoding in whatever way is appropriate.

The change in this case is a little broader than ideal, since it
interprets the entire header chunk (including the status line and
all headers) as ISO-8859-1. Mitigations for this could be put in
place if needed, but some more involved refactoring would likely be
required.

For what it's worth, this solution will make HTTP::UserAgent produce
similar results to Cro::HTTP::Client, and HTTP::Tiny. Curl-based Raku
libraries will, on the other hand, decode the duck as UTF-8 which,
although funnier, is arguably incorrect.

[RFC 7230 § 3.2.4]: https://tools.ietf.org/html/rfc7230#section-3.2.4

Fixes #236